添加令牌算法功能

+++ hmdp-redis-tool-framework/hmdp-redis-rate-limit-framework/src/main/java/org/javaup/lua/RateLimitOperate.java
@@ -31,7 +31,7 @@ public class RateLimitOperate {
         }
     }
     
-    public Integer execute(List<String> keys, String[] args){
-        return (Integer)redisCache.getInstance().execute(redisScript, keys, args);
+    public Long execute(List<String> keys, String[] args){
+        return (Long)redisCache.getInstance().execute(redisScript, keys, args);
     }
 }
